+++ jdaviz/core/loaders/importers/to_dc_with_label.vue
@@ -30,20 +30,14 @@
       hint="Select the viewer to use for the new data entry."
     ></plugin-viewer-create-new>
 
-    <v-row justify="end">
-      <plugin-action-button
-        :spinner="import_spinner"
-        :disabled="import_disabled"
-        :results_isolated_to_plugin="false"
-        :api_hints_enabled="api_hints_enabled"
-        @click="import_clicked">
-        {{ api_hints_enabled ?
-          'ldr.load()'
-          :
-          'Import'
-        }}
-      </plugin-action-button>
-    </v-row>
+    <loader-import-button
+      :spinner="import_spinner"
+      :disabled="import_disabled"
+      :api_hints_enabled="api_hints_enabled"
+      api_hint="ldr.load()"
+      :data_label_overwrite="data_label_overwrite"
+      @click="import_clicked">
+    </loader-import-button>
 
   </v-container>
 </template>